refactor: clearAuthState 함수 위치 변경 (#30)
