+++ src/services/utils/tokenManager.ts
@@ -59,3 +59,33 @@ export const clearTokens = () => {
 export const hasTokens = (): boolean => {
   return !!(getAccessToken() && getRefreshToken());
 };
+
+/**
+ * 전역 상태 및 토큰 클리어
+ */
+export const clearAuthState = async () => {
+  try {
+    // 1. 토큰 삭제
+    clearTokens();
+
+    // 2. React Query 캐시 클리어
+    if (typeof window !== 'undefined') {
+      const {getQueryClient} = await import('@/app/providers');
+      const queryClient = getQueryClient();
+      queryClient.clear();
+    }
+
+    // 3. 전역 상태 초기화
+    if (typeof window !== 'undefined') {
+      const {useAuthStore} = await import('@/store/useAuthStore');
+      useAuthStore.getState().logout();
+    }
+
+    // 4. 기본 페이지로 리다이렉트
+    if (typeof window !== 'undefined') {
+      window.location.href = '/';
+    }
+  } catch (error) {
+    console.error('Failed to clear auth state:', error);
+  }
+};
